#20 feat: add custom exceptions

+++ backend/service/MovieService.py
@@ -1,3 +1,4 @@
+from exception.MovieNotFoundException import MovieNotFoundException
 from repository.MovieRepository import repository
 from mapper.MovieMapper import MovieMapper as mapper
 from dto.MovieDTO import (
@@ -12,6 +13,9 @@ class MovieService:
         return [mapper.to_dto(dto_model=MovieDTO, orm_model=movie) for movie in movies]
     
     async def get_by_id(self, id: int) -> MovieDTO:
+        if await repository.get_by_id(id) is None:
+            raise MovieNotFoundException()
+
         movie = await repository.get_by_id(id)
         return mapper.to_dto(dto_model=MovieDTO, orm_model=movie)
     
@@ -21,11 +25,17 @@ class MovieService:
         return mapper.to_dto(dto_model=MovieDTO, orm_model=created_movie)
     
     async def update(self, id: int, dto: UpdateMovieDTO) -> MovieDTO:
+        if await repository.get_by_id(id) is None:
+            raise MovieNotFoundException()
+
         dict_movie = mapper.to_dict(dto)
         created_movie = await repository.update(id, dict_movie)
         return mapper.to_dto(dto_model=MovieDTO, orm_model=created_movie)
     
     async def delete(self, id: int) -> MovieDTO:
+        if await repository.get_by_id(id) is None:
+            raise MovieNotFoundException()
+
         movie = await repository.delete(id)
         return mapper.to_dto(dto_model=MovieDTO, orm_model=movie)
